Add files via upload
Question 1051. Height Checker.

Prompt: 
A school is trying to take an annual photo of all the students. The students are asked to stand in a single file line in non-decreasing order by height. Let this ordering be represented by the integer array expected where expected[i] is the expected height of the ith student in line.

You are given an integer array heights representing the current order that the students are standing in. Each heights[i] is the height of the ith student in line (0-indexed).

Return the number of indices where heights[i] != expected[i].

Here, the answer is simple. We need to create an exact duplicate of the heights we are given, then sort them in a descending order. The final step is to just compare the two arrays and keep a count of all the ones that are not equal.